refactor: split bff public type boundaries
