+++ apps/frontend-bff/tests/home-page-client.test.tsx
@@ -5,7 +5,7 @@ import { act } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import type { HomeResponse } from "../src/runtime-types";
+import type { HomeResponse } from "../src/public-types";
 
 const homeDataMocks = vi.hoisted(() => ({
   createWorkspaceFromHome: vi.fn(),
